api/requirements.txt
flask==1.0.2
pillow==5.0.0
numpy==1.13.3
tensorflow==1.8.0
tensorflow-serving-api-python3==1.8.0
uwsgi==2.0.17.1
voluptuous==0.11.5
